+++ py/torch_tensorrt/dynamo/backend/conversion.py
@@ -2,8 +2,8 @@ from typing import Sequence, Union
 import torch
 import io
 from torch_tensorrt.fx.trt_module import TRTModule
-from torch_tensorrt.dynamo.backend._settings import CompilationSettings
-from torch_tensorrt.dynamo.fx_ts_compat.fx2trt import (
+from torch_tensorrt.dynamo.common import (
+    CompilationSettings,
     InputTensorSpec,
     TRTInterpreter,
 )
