+++ py/torch_tensorrt/dynamo/backend/utils.py
@@ -2,10 +2,9 @@ import torch
 import logging
 from dataclasses import replace, fields
 
-from torch_tensorrt.dynamo.backend._settings import CompilationSettings
+from torch_tensorrt.dynamo.common import CompilationSettings, use_python_runtime_parser
 from typing import Any, Union, Sequence, Dict
 from torch_tensorrt import _Input, Device
-from ..common_utils import use_python_runtime_parser
 
 
 logger = logging.getLogger(__name__)
