+++ py/torch_tensorrt/dynamo/fx_ts_compat/__init__.py
@@ -6,8 +6,6 @@ from torch_tensorrt.fx.converter_registry import (  # noqa
     NO_IMPLICIT_BATCH_DIM_SUPPORT,
     tensorrt_converter,
 )
-from .fx2trt import TRTInterpreter, TRTInterpreterResult  # noqa
-from .input_tensor_spec import InputTensorSpec  # noqa
 from .lower_setting import LowerSetting  # noqa
 from .lower import compile  # usort: skip  #noqa
