+++ py/torch_tensorrt/dynamo/test/test_dynamo_backend.py
@@ -8,7 +8,7 @@ import torchvision.models as models
 
 from transformers import BertModel
 
-from torch_tensorrt.dynamo.common_utils.test_utils import (
+from torch_tensorrt.dynamo.common.test_utils import (
     COSINE_THRESHOLD,
     cosine_similarity,
 )
